+++ js/applicationmenu.js
@@ -34,7 +34,9 @@
     function ApplicationMenu(element) {
       this.settings = $.extend({}, settings);
       this.element = $(element);
+      Soho.logTimeStart(pluginName);
       this.init();
+      Soho.logTimeEnd(pluginName);
     }
 
     // Plugin Methods