+++ js/environment.js
@@ -104,6 +104,33 @@
         $.detectBaseTag();
       }
 
+      window.Soho.logTimeStart = function(label) { // jshint ignore:line
+        if (window.Soho.logTime) {
+          console.time(label); // jshint ignore:line
+        }
+      };
+
+      window.Soho.logTimeEnd = function(label) { // jshint ignore:line
+        if (window.Soho.logTime) {
+          console.timeEnd(label); // jshint ignore:line
+        }
+      };
+
+      //Handle Personalization and theme
+      window.Soho.theme = 'light';
+
+      window.Soho.setTheme = function(theme) { // jshint ignore:line
+        $('body').trigger('changetheme', theme);
+      };
+
+      window.Soho.setColors = function(colors) { // jshint ignore:line
+        $('body').trigger('changecolors', colors);
+      };
+
+      $(function() {
+        $('body').personalize();
+      });
+
       return this;
     }