+++ js/lightbox.js
@@ -43,7 +43,9 @@
     function Lightbox(element) {
       this.settings = $.extend({}, settings);
       this.element = $(element);
+      Soho.logTimeStart(pluginName);
       this.init();
+      Soho.logTimeEnd(pluginName);
     }
 
     // Lightbox Methods