+++ js/mask.js
@@ -52,7 +52,9 @@
     function Mask(element) {
       this.settings = $.extend({}, settings);
       this.element = $(element);
+      Soho.logTimeStart(pluginName);
       this.init();
+      Soho.logTimeEnd(pluginName);
     }
 
     Mask.prototype = {