+++ js/slider.js
@@ -38,7 +38,9 @@
     function Slider(element) {
       this.settings = $.extend({}, settings);
       this.element = $(element);
+      Soho.logTimeStart(pluginName);
       this.init();
+      Soho.logTimeEnd(pluginName);
     }
 
     // Check if is an integer
@@ -779,9 +781,9 @@
       },
 
       getColorClosestToValue: function() {
-        var currentTheme = $('#sohoxi-stylesheet').get(0).href.replace(/^.*[\\\/]/, '').replace(/\.[^\.]+$/, ''),
+        var currentTheme = Soho.theme,
           preColors = {
-            'grey-theme': {
+            'light': {
               'default'   : '#000000',
               'very-poor' : '#a13030',
               'poor'      : '#d66221',
@@ -790,7 +792,7 @@
               'very-good' : '#76b051',
               'superior'  : '#488421'
             },
-            'dark-theme': {
+            'dark': {
               'default'   : '#ffffff',
               'very-poor' : '#a13030',
               'poor'      : '#d66221',
@@ -799,7 +801,7 @@
               'very-good' : '#76b051',
               'superior'  : '#488421'
             },
-            'high-contrast-theme': {
+            'high-contrast': {
               'default'   : '#000000',
               'very-poor' : '#a13030',
               'poor'      : '#d66221',
@@ -809,12 +811,6 @@
               'superior'  : '#397514'
             }
           };
-
-        //remove from ? to end
-        var idx = currentTheme.indexOf('?');
-        if (currentTheme !== '' && idx > -1) {
-          currentTheme = currentTheme.substr(0, idx).replace('.css', '');
-        }
 
         var themeColors = preColors[currentTheme],
           val = this.value()[0],