+++ js/spinbox.js
@@ -28,7 +28,9 @@
     // Plugin Constructor
     function Spinbox(element) {
       this.element = $(element);
+      Soho.logTimeStart(pluginName);
       this.init();
+      Soho.logTimeEnd(pluginName);
     }
 
     // Plugin Methods